requirements-dev.txt
-e .
aiohttp==3.8.1
aiomcache==0.7.0
aioredis==2.0.0
attrs==21.4.0
chardet==4.0.0
cryptography==36.0.1
docker==5.0.3
flake8==4.0.1
flake8-bandit==2.1.2
flake8-bugbear==22.1.11
flake8-import-order==0.18.1
flake8-requirements==1.5.2
multidict==6.0.2
mypy==0.930
pep257==0.7.0
pre-commit==2.17.0
pynacl==1.5.0
pytest==6.2.5
pytest-aiohttp==1.0.3
pytest-cov==3.0.0
pytest-mock==3.6.1
pytest-sugar==0.9.4
sphinx==4.3.2
typing-extensions==4.0.1
yarl==1.7.2
